pyproject.toml
[tool.poetry]
name = "schedule"
version = "0.1.0"
description = ""
authors = ["lifeissea <lifeissea@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
